Changed mortgage calculator to use doubles instead of floats to get a more accurate answer

+++ src/main/java/com/pluralsight/MortgageCalculator.java
@@ -6,14 +6,14 @@ public class MortgageCalculator {
     public static void calculate(Scanner scanner) {
         System.out.println("\n\nWelcome to the Mortgage calculator!\n\n");
         System.out.println("Please enter your loan amount: ");
-        float loanAmount = scanner.nextFloat();
+        double loanAmount = scanner.nextFloat();
         System.out.println("Please enter your current interest rate per year: ");
-        float interestRate = scanner.nextFloat();
+        double interestRate = scanner.nextFloat();
         System.out.println("Please enter your payment length in years: ");
         int loanLength = scanner.nextInt();
 
         int n = loanLength * 12;
-        float i = (interestRate / 100) / 12;
+        double i = (interestRate / 100) / 12;
         double monthlyPayment = (loanAmount * (i * Math.pow((1 + i), n))) / (Math.pow((1 + i), n) - 1);
         double totalInterest = (monthlyPayment * n) - loanAmount;
